Create thorski-request
Hello - 
I am the marketing manager for Thorski Design (thorski.com) and noticed you pulled a copy of Thorski's website. We would like to request that you remove this repo.

The images on Thorski.com are copyrighted and are not for public use. 

Thank you,
Sincerely,

Lisa Hulen on behalf of 
Eric Thorstenson
Thorski Design, LLC